Fixes: eliminating owner type parameters as we no longer need that. Also, eliminating other warnings.

+++ src/share/classes/org/paninij/lang/Long.java
@@ -493,7 +493,7 @@ public class Long extends Number implements Comparable<Long> {
      * {@code long} value.
      */
     public long longValue() {
-        return (long)value;
+        return value;
     }
 
     /**
